modules/location_consumer/requirements.txt
kafka-python
GeoAlchemy2==0.12.3
psycopg2-binary==2.9.3
shapely==1.8.2